Session 2
During this session I managed to add extending slides for our robot (they are made by gobilda and the files were imported), also when I added them, there were clearance problems between the slide motor and the drivetrain motor so I had to adjust the mounting holes of the motors in all plates, as well as completely redesign the plate to which the motors are attached with screws. Additionally, I added control and expansion hubs onto the robot, which are required to control the robot. Moreover, I added mounting points for the intake, as well as transfer mechanism onto the robot and connected both sides of the drivetrain together. Now I will need to add odometry pods and create mounting blocks for the slides, which I will need to 3d print