Session 3
I made the mounts that I will be 3d printing and using threaded brass heat pressed inserts to mount to connect the slides to the side inner walls of the drivetrain. Additionally, I mounted odometry pods, which I will be using for autonomous robot driving, as well as to properly use driver-oriented driving.