Session 4
During this session I designed a mount for our battery, which has a guide for wires do go to the bottom of the robot without getting hit by anything in the middle of the robot. Also, this holder has a lot of holes, which should help with heat dissipation, as during use our batteries tend to get very hot.