Session 7
During this session I managed to almost finish making the outtake mechanism for my FTC Robot. I made all parts needed for the basket itself, but I still need to make a way to mount the servo and add the screws, as I have previous experience of not doing so and then having many clearance issues when the parts are already manufactured, so that is what I plan do to next.