Session 8
During this session I managed to design the mounts for SG90 servo motors, successfully added all screws into the robot and already noticed one big problem, where the U-Channel connecting both inner sides of the robot (you can see the problem in the last photo). Also, I managed design the mounts for the cover of the robot from the front, which should make the robot more robust as well as, to prevent things that are not supposed to get into the robot.