Session 9
I managed to make the side walls of the intake, as well as added bearings and shafts that go across the intake and will be responsible for pushing game elements towards the outtake. I still need to make a plate for the game pieces to be pushed on as well as, to add surgical tubing brushes, that will be pushing those elements, however the hardest part is done if the geometry works as expected.