Session 10
During this session I managed to design roller for our intake mechanism but instead of using a surgical tubing one, I will try to experiment with one printed out of TPU. Also I made a plate and walls for it, on which pixels should be able to slide towards the outtake of the robot. Additionally, I made a counter roller, added gears for it and added lifting part of the intake that will make our intake be able to reach further and into harder spaces to drive into.