Session 11
During this session I managed to complete a large part of the intake. I placed "optake" surgical tubing brush holders in their spaces and added spacers to first and second row of brushes. More importantly, I designed a pulley integrated into the intake, which will allow me to lift the intake up and down using a servo motor and a timing belt and this is extremely important, as it will make the robot smaller, which makes it easier to drive as well as makes it less likely that it will break during a match.