Session 13
During this session I mounted the motor on the intake and made a plate that will guide the game elements towards the outtake. Now I have 2 important things to make. First of them is a guard that will prevent game elements from getting stuck in the gears and another one is that I need to change the angle of one of the intake guide-walls as currently two game elements could get stuck if they would be picked up at the same time.