Session 14
I managed to make a guard that should prevent game elements from getting into the gears. Also, I was not sure about the correct angle needed for the game elements to slide past each other without getting stuck, so I made a prototype out of plywood that allows me to test different angles of the sides to find out which ones work the best and after that I modified the angle of guide walls in my robot file.