Session 18
I removed a bit of the plate that holds the motor, as it was stopping the intake from lifting up all the way (last picture). Also, I shifted the plate on which the pixels (game elements) slide down by a few millimeters, so they would not get stuck between the counter-roller and the ramp itself. Next I will have to remanufacture the intake to test for new bugs and I think I might try to make the drivetrain or will finish making a drone launcher (paper airplane launcher) or a mechanism that is required for the autonomous mode.