Session 19
I removed the top part of the robot to minimize the waste of the material and added mounts on which I will be able to mount the control units for the robot as well as the battery. Also, I added mounting on which I will mount a storage, which will store the game elements collected during testing, as there is no way to take them out. Tomorrow I will go to the public workshop and will try to laser cut the plates for the updated version of the intake, as well as the drivetrain and after I will have all 3d printed and laser cut parts I will try to assemble everything for testing.